+++ app/assets/javascripts/server_info.js
@@ -2,15 +2,20 @@
  * Created by Pavel.Lobashov on 14.10.15.
  */
 fetch_server_ip = function (server) {
-    var ip;
     $.ajax({
         url: '/servers/cloud_server_fetch_ip',
         type: 'GET',
         data: {
             'server': server
         },
+        beforeSend: function () {
+            showOverlay('Fetching current server IP');
+        },
+        complete: function () {
+            hideOverlay();
+        },
         success: function (data) {
-            update_ip_value(data['ip']);
+            update_ip_value(data.ip);
         },
         error: function (e) {
             console.log(e.message);
@@ -22,6 +27,6 @@ update_ip_value = function(value) {
         $("#server_address").val(value);
 };
 
-$("#fetch-ip").on("click", function(){
+function fetchCurrentServerIp() {
     fetch_server_ip($( "#server_name" ).val());
-});
+}